backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==3.14.0
django-filter==23.1
djangorestframework-simplejwt==5.2.2
mysqlclient==2.1.1
python-dotenv==1.0.0
celery==5.2.7
redis==4.5.4
pycryptodome==3.17
bcrypt==4.0.1
gunicorn==20.1.0
PyJWT==2.6.0
drf-yasg==1.21.5
pytest==7.3.1
flake8==6.0.0
